requirements.txt
openai~=1.60.2
torch~=2.6.0
nltk~=3.9.1
transformers~=4.49.0
numpy~=1.26.4
python-dotenv~=1.0.1
pytest~=8.3.4